requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings
python-dotenv==1.0.0


# Database
sqlalchemy==2.0.35
alembic==1.13.0
asyncpg==0.30.0        


# Redis & Celery (для очередей задач)
redis==5.0.0
celery==5.4.0

# Yandex Cloud AI
yandex-cloud-ml-sdk==0.15.0   
openai


# Аналитика 
pandas==2.2.0
numpy==1.26.0

# HTTP клиенты
httpx==0.27.0

email-validator

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
bcrypt==4.0.1